requirements.txt
google-adk[database]==0.3.0
google-generativeai==0.8.5
python-dotenv==1.1.0
requests==2.31.0
pandas

# FastAPI server dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Google Cloud Services for Hackathon
googlemaps==4.10.0
google-cloud-translate==3.15.3
google-api-python-client>=2.157.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-vision==3.7.2
google-cloud-speech==2.26.0
google-cloud-texttospeech==2.16.3

# Weather API
pyowm==3.3.0
